add-message

+++ SocialNetwork.Business/Services/Implements/PostCommentService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Identity;
 using SocialNetwork.Business.Constants;
 using SocialNetwork.Business.DTOs.CommentReaction.Requests;
 using SocialNetwork.Business.DTOs.CommentReaction.Responses;
@@ -9,29 +10,32 @@ using SocialNetwork.Business.Wrapper;
 using SocialNetwork.Business.Wrapper.Interfaces;
 using SocialNetwork.DataAccess.Entities;
 using SocialNetwork.DataAccess.Repositories.Interfaces;
+using SocialNetwork.DataAccess.Utilities.Roles;
 
 namespace SocialNetwork.Business.Services.Implements
 {
     public class PostCommentService : BaseServices, IPostCommentService
     {
-        public PostCommentService(IUnitOfWork unitOfWork, IMapper mapper) : base(unitOfWork, mapper)
+        private readonly UserManager<User> _userManager;
+        public PostCommentService(IUnitOfWork unitOfWork, IMapper mapper, UserManager<User> userManager) : base(unitOfWork, mapper)
         {
+            _userManager = userManager;
         }
 
         #region Comment
-        public async Task<IResponse> Create(CreatePostCommentRequest request)
+        public async Task<IResponse> Create(string requestUserId, CreatePostCommentRequest request)
         {
-            if (!await CheckExitsPost(request.PostId))
-            {
-                return new ErrorResponse(404, Messages.NotFounb("Post"));
-            }
+            var post = await _unitOfWork.PostRepository.GetById(request.PostId);
+            if (post == null) { return new ErrorResponse(404, Messages.NotFounb("Post")); }
 
-            if (!await CheckExitsUser(request.UserId))
+            if (!await CheckAccessPost(requestUserId, post.AuthorId))
             {
-                return new ErrorResponse(404, Messages.NotFounb("User"));
-            }
+                return new ErrorResponse(400, Messages.NotFriend);
+            }    
 
             var addComment = _mapper.Map<PostComment>(request);
+            addComment.UserId = requestUserId;
+
             await _unitOfWork.PostCommentRepository.Add(addComment);
 
             var result = await _unitOfWork.CompleteAsync();
@@ -44,8 +48,20 @@ namespace SocialNetwork.Business.Services.Implements
             return new DataResponse(_mapper.Map<GetPostCommentReponse>(addComment), 200, Messages.CreatedSuccessfully);
         }
 
-        public async Task<IResponse> Delete(Guid id)
+        public async Task<IResponse> Delete(string requestUserId, Guid id)
         {
+            var comment = await _unitOfWork.PostCommentRepository.GetById(id);
+
+            if (comment == null)
+            {
+                return new ErrorResponse(404, Messages.NotFound);
+            }    
+
+            if (comment.UserId != requestUserId)
+            {
+                return new ErrorResponse(400, Messages.CommentNotOwner);
+            }    
+
             if (!await _unitOfWork.PostCommentRepository.Delete(id))
             {
                 return new ErrorResponse(404, Messages.NotFounb("Comment"));
@@ -67,27 +83,39 @@ namespace SocialNetwork.Business.Services.Implements
             return new DataResponse(_mapper.Map<List<GetPostCommentReponse>>(entity), 200);
         }
 
-        public async Task<IResponse> GetById(Guid id)
+        public async Task<IResponse> GetById(string requestUserId, Guid id)
         {
-            var entity = await _unitOfWork.PostCommentRepository.GetById(id);
-            if (entity == null)
+            var comment = await _unitOfWork.PostCommentRepository.GetById(id);
+
+            if (comment == null)
             {
                 return new ErrorResponse(404, Messages.NotFound);
             }    
 
-            return new DataResponse(_mapper.Map<GetPostCommentReponse>(entity), 200);
+            if (!await CheckAccessPost(requestUserId, comment.UserId))
+            {
+                return new ErrorResponse(400, Messages.NotFriend);
+            }    
+
+            return new DataResponse(_mapper.Map<GetPostCommentReponse>(comment), 200);
 
         }
 
-        public async Task<IResponse> Update(Guid id, UpdatePostCommentRequest request)
+        public async Task<IResponse> Update(string requestUserId, Guid id, UpdatePostCommentRequest request)
         {
-            var updateEntity = _mapper.Map<PostComment>(request);
-            updateEntity.Id = id;
+            var comment = await _unitOfWork.PostCommentRepository.GetById(id, false);
 
-            if (!await _unitOfWork.PostCommentRepository.Update(updateEntity))
+            if (comment == null) 
             {
-                return new ErrorResponse(404, Messages.NotFounb("Comment"));
+                return new ErrorResponse(404, Messages.NotFound);
             }
+
+            if (!await CheckOwnerComment(requestUserId, comment.UserId))
+            {
+                return new ErrorResponse(400, Messages.CommentNotOwner);
+            }
+
+            _mapper.Map(request, comment);
 
             var result = await _unitOfWork.CompleteAsync();
             
@@ -96,30 +124,75 @@ namespace SocialNetwork.Business.Services.Implements
                 return new ErrorResponse(400, Messages.UpdateError);
             }
 
-            return new DataResponse(_mapper.Map<GetPostCommentReponse>(updateEntity), 204, Messages.UpdatedSuccessfully);
+            return new DataResponse(_mapper.Map<GetPostCommentReponse>(comment), 204, Messages.UpdatedSuccessfully);
 
         }
-    
-        private async Task<bool> CheckExitsUser(string userId)
+
+        private async Task<bool> CheckAccessPost(string requestUserId, string targetUserId)
         {
-            var user = await _unitOfWork.UserRepository.FindById(userId);
+            var requestUser = await _unitOfWork.UserRepository.FindById(targetUserId);
+            var targetUser = await _unitOfWork.UserRepository.FindById(requestUserId);
 
-            return user != null;
+            if (requestUser == null || targetUser == null) return false;
+
+            if (requestUserId == targetUserId)
+            {
+                return true;
+            }
+
+            // Is admin
+            if (await _userManager.IsInRoleAsync(targetUser, RoleName.Administrator))
+            {
+                return true;
+            }
+
+            // Is friend
+            return await _unitOfWork.FriendshipRepository.IsFriend(requestUserId, targetUserId);
         }
-
-        private async Task<bool> CheckExitsPost(Guid Id)
+        private async Task<bool> CheckOwnerComment(string requestUserId, string authorCommentId)
         {
-            var post = await _unitOfWork.PostRepository.GetById(Id);
+            var requestUser = await _unitOfWork.UserRepository.FindById(requestUserId);
+            if (requestUser == null) return false;
 
-            return post != null;
+            // Is owner comment
+            if (requestUserId == authorCommentId)
+            {
+                return true;
+            }
+
+            // Is admin
+            if (await _userManager.IsInRoleAsync(requestUser, RoleName.Administrator))
+            {
+                return true;
+            }
+
+            return false;
         }
+
 
         #endregion
 
         #region Comment Reaction
 
-        public async Task<IResponse> GetReactions(Guid commentId)
+        public async Task<IResponse> GetReactions(string requestUserId, Guid commentId)
         {
+            var comment = await _unitOfWork.PostCommentRepository.GetById(commentId);
+            if (comment == null)
+            {
+                return new ErrorResponse(404, Messages.NotFound);
+            }
+
+            var post = await _unitOfWork.PostRepository.GetById(comment.PostId);
+            if (post == null)
+            {
+                return new ErrorResponse(404, Messages.NotFounb("Post of comment"));
+            }    
+
+            if (!await CheckAccessPost(requestUserId, post.AuthorId))
+            {
+                return new ErrorResponse(404, Messages.NotFriend);
+            }
+
             var result = await _unitOfWork.CommentReactionRepository.GetByComment(commentId);
 
             return new DataResponse(_mapper.Map<List<GetCommentReactionResponse>>(result), 200);
@@ -215,6 +288,7 @@ namespace SocialNetwork.Business.Services.Implements
         }
 
         #endregion
+
     }
 
 }
